requirements.txt
blinker==1.6.2
boto3==1.26.149
botocore==1.29.149
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
flake8==6.0.0
Flask==2.3.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector-python==8.0.33
packaging==23.1
pluggy==1.0.0
protobuf==3.20.3
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.5
